requirements.txt
playwright
beautifulsoup4
google-generativeai
python-dotenv
chromadb
sentence-transformers
pandas